We record liquefiables into the layer for all API levels now, so fixed this bug.
- Also refactored some of the logic to keep the renderEffect private in LiquidScopeImpl.
