+++ liquid/src/main/kotlin/io/github/fletchmckee/liquid/internal/LiquidBackupElement.kt
@@ -4,8 +4,15 @@ package io.github.fletchmckee.liquid.internal
 
 import android.os.Build
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.RectangleShape
+import androidx.compose.ui.graphics.drawOutline
 import androidx.compose.ui.graphics.drawscope.ContentDrawScope
+import androidx.compose.ui.graphics.drawscope.Fill
+import androidx.compose.ui.graphics.drawscope.clipPath
+import androidx.compose.ui.graphics.drawscope.translate
+import androidx.compose.ui.graphics.isSpecified
 import androidx.compose.ui.graphics.layer.GraphicsLayer
+import androidx.compose.ui.graphics.layer.drawLayer
 import androidx.compose.ui.layout.LayoutCoordinates
 import androidx.compose.ui.layout.positionOnScreen
 import androidx.compose.ui.node.CompositionLocalConsumerModifierNode
@@ -99,8 +106,6 @@ internal class LiquidBackupNode(
     }
 
     if (shouldInvalidate) {
-      // Clear the blur effect if frost changed (only relevant for API 31+)
-      reusableScope.renderEffect = reusableScope.renderEffect.takeUnless { canUseRenderEffect && reusableScope.mutatedFields has Fields.Frost }
       invalidateDraw()
     }
   }
@@ -131,7 +136,37 @@ internal class LiquidBackupNode(
     try {
       val layer = obtainGraphicsLayer()
       reusableScope.density = currentValueOf(LocalDensity)
-      drawBackupLiquidEffect(layer, reusableScope)
+      val shapeOutline = reusableScope.shape.createOutline(size, layoutDirection, this)
+      val shapePath = shapeOutline.asPath()
+      val frostRadius = reusableScope.frostRadius
+      // We still record liquefiables into the layer for Android 11 and lower.
+      // Otherwise something like a shadow effect for a transparent liquid composable will appear inside the node.
+      recordLiquefiablesIntoLayer(layer, reusableScope)
+      layer.clip = reusableScope.shape != RectangleShape
+
+      if (frostRadius > 0f && Build.VERSION.SDK_INT >= 31) {
+        // If we have a valid frostRadius and the device is API 31 or 32, we can at least use Android's BlurEffect.
+        layer.renderEffect = reusableScope.obtainPreTiramisuRenderEffect()
+      }
+
+      clipPath(shapePath) {
+        translate(-frostRadius, -frostRadius) { drawLayer(layer) }
+      }
+
+      // Fill the shape with the tint if one is provided.
+      if (reusableScope.tint.isSpecified) {
+        drawOutline(
+          outline = shapeOutline,
+          color = reusableScope.tint,
+          style = Fill,
+        )
+      }
+
+      if (reusableScope.edge > 0f) {
+        drawBackupEdgeEffect(shapePath)
+      }
+
+      drawContent()
     } finally {
       reusableScope.reset()
     }
